requirements.txt
openai>=1.0.0
playwright>=1.35.0
sounddevice>=0.4.5
soundfile>=0.12.1
python-dotenv>=1.0.0
pyttsx3>=2.90
requests>=2.28.0
